Inline the "Prerequisite" Chapter
This will make the documentation easier to understand, because topics such as db, oauth, domain, registry, certs., etc. can now be covered completely by their own chapter.

This also allows to focus on the relevant prerequisites for the environment of choice. For example, on GCP the build-in docker registry is not relevant.